Single Code Base for Mobile and Desktop Versions
Earlier there used to be two different branches for developing two different versions of Desktop and Mobile.
v1.x series was used to be for Desktop and v2.x was used for Android.

From now onwards, v3.x will be single code base for both type of platforms.
This step will allow us to provide rapid development of Matrix Calculator.